Refactor bash tool
