+++ src/tools/run_python_tool.py
@@ -1,16 +1,15 @@
 from pydantic_ai import Tool
 from pathlib import Path
-from .bash import ExclusiveBashProcess
+from .bash_tool import BashProcess
 
-def create_run_python_tool(agent_id, workspace_dir, run_mode, timeout, max_retries, proxy, conda_prefix=True):
-    bash = ExclusiveBashProcess(
+def create_run_python_tool(agent_id, workspace_dir, run_mode, timeout, max_retries, proxy):
+    bash = BashProcess(
         agent_id=agent_id,
         workspace_dir=workspace_dir,
         run_mode=run_mode,
         autoconda=False,
         timeout=timeout,
         proxy = proxy,
-        auto_torch=False,
     )
 
     def _run_python(python_file_path: str):
@@ -28,13 +27,10 @@ def create_run_python_tool(agent_id, workspace_dir, run_mode, timeout, max_retri
             return "python_file_path is not a valid python file path"
         
         #TODO allow to accept arguments + validate they don't break the bash (requiring input etc)
-        if(conda_prefix):
-            env_path = workspace_dir / agent_id / ".conda" / "envs" / f"{agent_id}_env"
-            command_prefix=f"source /opt/conda/etc/profile.d/conda.sh && conda activate {env_path} && "
-            command = command_prefix + f"python {python_file_path}"
+        env_path = workspace_dir / agent_id / ".conda" / "envs" / f"{agent_id}_env"
+        command = f"conda run -p {env_path} --no-capture-output python {python_file_path}"
         out = bash.run(command)
-        if(len(out) > 5000):
-            out = out[:5000]+"\n ... (output truncated, too long)"
+        
         return out
     
     run_python_tool = Tool(
